Prepare for tests.
