+++ Package.swift
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
   name: "SwiftDGC",
   defaultLocalization: "en",
-  platforms: [.iOS(.v12)],
+  platforms: [.iOS(.v12), .macOS(.v10_14)],
   products: [
     // Products define the executables and libraries a package produces, and make them visible to other packages.
     .library(
